requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-dotenv>=1.0.0
supabase==2.8.0
openai>=1.3.0
pydantic>=2.4.0
httpx==0.27.0
